Fixed usage of `int`.

+++ PaymentKit/PKCardNumber.m
@@ -37,7 +37,7 @@
 
     NSString *firstChars = [_number substringWithRange:NSMakeRange(0, 2)];
 
-    int range = [firstChars integerValue];
+    NSInteger range = [firstChars integerValue];
 
     if (range >= 40 && range <= 49) {
         return PKCardTypeVisa;
